+++ src/constants.ts
@@ -22,20 +22,3 @@ export const SIDE_PANEL_WIDTH: number = 280;
 export const MIN_WIDTH_TO_SHOW_SIDE_PANEL: number = 580;
 export const VIEWER_INITIAL_WIDTH: number = 500;
 export const VIEWER_HEIGHT: number = 529;
-
-export interface TimeUnits {
-  magnitude: number;
-  name: string;
-}
-
-export interface PlaybackState {
-  currentTime: number;
-  isPlaying: boolean;
-}
-
-export interface PlaybackData {
-  timeStep: number;
-  lastFrameTime: number;
-  firstFrameTime: number;
-  timeUnits: TimeUnits;
-}
